pyproject.toml
[tool.poetry]
name = "merrycrispr"
version = "3.0.1"
description = "Generate gene and genome-wide CRISPR/Cas9 knockout and excision libraries"
authors = ["milescsmith <mileschristiansmith@gmail.com>"]
license = "LGPL-3.0-or-later"
homepage = "https://github.com/milescsmith/merrycrispr"
repository = "https://github.com/milescsmith/merrycrispr"
readme = "README.rst"
classifiers = [  # Update me
    "Development Status :: 3 - Production",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: POSIX :: Linux",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
keywords = ["CRISPR"]
include = ["data/*.*"]

[tool.poetry.scripts]
"merrycrispr" = "merrycrispr.__main__:main"

[tool.poetry.dependencies]
python = "^3.9,<3.10"
gtfparse = { git = "https://github.com/milescsmith/gtfparse.git", branch = "main" }
biopython = "^1.79"
regex = "^2022.3.15"
pyfaidx = "^0.6.4"
click = "^8.0.4"
pandas = "^1.4.1"
requests = "^2.27.1"
p-tqdm = "^1.3.3"
tqdm = "^4.63.0"
azimuth = {git = "https://github.com/milescsmith/azimuth", branch = "main"}

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"
mypy = "^0.941"
flake8 = "^4.0.1"
pyre-check = "^0.9.10"
Sphinx = "^4.4.0"
sphinx-rtd-theme = "^1.0.0"
sphinxcontrib-applehelp = "^1.0.2"
sphinxcontrib-devhelp = "^1.0.2"
sphinxcontrib-jsmath = "^1.0.1"
sphinxcontrib-qthelp = "^1.0.3"
sphinxcontrib-serializinghtml = "^1.1.5"
ipython = "^8.1.1"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
